refactor: split panic handler code into panic.rs
