+++ rust-version/src/main.rs
@@ -5,13 +5,12 @@
 #![no_main]
 #![no_std]
 
-use aarch64_cpu::asm;
 use core::arch::global_asm;
-use core::panic::PanicInfo;
 
 mod console;
 mod print;
 mod synchronization;
+mod panic;
 
 //--------------------------------------------------------------------------------------------------
 // Public Code
@@ -52,58 +51,3 @@ global_asm!(include_str!("boot.S"),
 CONST_CORE_ID_MASK = const 0b11
 );
 
-/// Stop immediately if called a second time.
-///
-/// # Note
-///
-/// Using atomics here relieves us from needing to use `unsafe` for the static variable.
-///
-/// On `AArch64`, which is the only implemented architecture at the time of writing this,
-/// [`AtomicBool::load`] and [`AtomicBool::store`] are lowered to ordinary load and store
-/// instructions. They are therefore safe to use even with MMU + caching deactivated.
-///
-/// [`AtomicBool::load`]: core::sync::atomic::AtomicBool::load
-/// [`AtomicBool::store`]: core::sync::atomic::AtomicBool::store
-fn panic_prevent_reenter() {
-    use core::sync::atomic::{AtomicBool, Ordering};
-
-    static PANIC_IN_PROGRESS: AtomicBool = AtomicBool::new(false);
-
-    if !PANIC_IN_PROGRESS.load(Ordering::Relaxed) {
-        PANIC_IN_PROGRESS.store(true, Ordering::Relaxed);
-
-        return;
-    }
-
-    wait_forever()
-}
-
-#[panic_handler]
-fn panic(info: &PanicInfo) -> ! {
-    // Protect against panic infinite loops if any of the following code panics itself.
-    panic_prevent_reenter();
-
-    let (location, line, column) = match info.location() {
-        Some(loc) => (loc.file(), loc.line(), loc.column()),
-        _ => ("???", 0, 0),
-    };
-
-    println!(
-        "Kernel panic!\n\n\
-        Panic location:\n      File '{}', line {}, column {}\n\n\
-        {}",
-        location,
-        line,
-        column,
-        info.message().unwrap_or(&format_args!("")),
-    );
-
-    wait_forever();
-}
-
-#[inline(always)]
-pub fn wait_forever() -> ! {
-    loop {
-        asm::wfe()
-    }
-}
